Write tests for validator API
