+++ packages/airnode-validator/src/v2/config/config.test.ts
@@ -3,6 +3,8 @@ import { join } from 'path';
 import { configSchema } from './config';
 
 it('successfully parses config.json specs', () => {
-  const ois = JSON.parse(readFileSync(join(__dirname, '../../../exampleSpecs/config.specs.json')).toString());
+  const ois = JSON.parse(
+    readFileSync(join(__dirname, '../../../exampleSpecs/interpolated-config.specs.json')).toString()
+  );
   expect(() => configSchema.parse(ois)).not.toThrow();
 });
